A simple dependency injection container.
This package provides the following
- A container where dependencies can be registered
- A wrapped container, allowing for chained lookups
- A method for injecting dependencies using the container
- A helper method for registering dependencies as interfaces

The core of this package is `Container`. It is created via the `New`
method. The container can provide dependencies statically or lazily.

The `Inject` method does the actual DI work. Dependencies into a
function by inspecting its signature and provides the arguments based
on their type.